requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.11.4
python-multipart==0.0.18
